% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbev_module.R
\name{comp_quant}
\alias{comp_quant}
\title{Compute quantiles and goodness of fit measures for a fitted model}
\usage{
comp_quant(model_fit, maxval, trmin = 2)
}
\arguments{
\item{model_fit}{list as returned from the fit_ev_model function}

\item{maxval}{array of annual maxima to be used for
model validation and for computing quantiles}

\item{trmin}{minimum return time for computing quantiles,
and to be included in the fractional square error / mean bias measures
(default trmin = 2)}
}
\value{
a list with the following named quantities
\describe{
  \item{qmean}{expected value for the computed quantiles (for Tr >= trmin)}
  \item{qupper}{upper credidility interval for quantiles (prob = 0.95)
  (quantiles only for Tr >= trmin)}
  \item{qlower}{lower credibility interval for quantiles (prob = 0.05)
  (quantiles only for Tr >= trmin)}
  \item{qmedian}{median value of the computed quantiles  (prob = 0.50)
  (quantiles only for Tr >= trmin)}
  \item{fmean}
  \item{fupper}
  \item{flower}
  \item{fmedian}
  \item{dmean}
  \item{dupper}
  \item{dlower}
  \item{dmedian}
  \item{lpml}
  \item{lppd}
  \item{fse}
  \item{mbias}
  \item{mwidth}
  \item{fse_Tr}
  \item{bias_Tr}
  \item{wisth_Tr}
  \item{elpd_loo}
  \item{p_loo}
  \item{elpd_waic1}
  \item{p_waic1}
  \item{elpd_waic2}
  \item{p_waic2}
  \item{quants}
  \item{cdfs}
  \item{pdfs}
  \item{epsi}
  \item{Fival}
  \item{Trval}
  \item{Xival}
  \item{FivalQ}
  \item{TrvalQ}
  \item{XivalQ}
  \item{log_lik}{log likelihood (nyears*ndraws)}
  \item{model}{name of the model}
  \item{Mgen}{number of draws for latent variable level for hbev models}
  \item{thresh_hbev}{theshold used for the hbev-family models}
  \item{nwarning_x0}{number of warnings for optimization convergence problems}
  }
}
\description{
Compute quantiles and goodness of fit measures for a fitted model
}
