% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbev_module.R
\name{fit_ev_model}
\alias{fit_ev_model}
\title{Fit an extreme value model to data}
\usage{
fit_ev_model(
  data,
  model = "wei_dyn_bin",
  Mgen = 50,
  thresh_pot = 10,
  thresh_hbev = 0,
  pot_frac = TRUE,
  frac_exc = 0.05,
  iter = 1000,
  chains = 4,
  onlymax_gev = FALSE,
  refresh = 0,
  empirical_prior = FALSE,
  adapt_delta = 0.8,
  priorpar = NULL,
  draw_priors = FALSE
)
}
\arguments{
\item{data}{matrix with daily precipitation data, with size
(number of blocks)*(number of observations in each block)
so that each row is the sample within one block.
Alternatively, it can be a vector with the block maxima if one sets
the parameter onlymax_gev = TRUE (only for the GEV model).}

\item{model}{which extreme value model.
Default is model='wei_dyn_bin'}

\item{Mgen}{Number of samples to draw for the latent level variable models
( for models of the HBEV family only)}

\item{thresh_pot}{threshold for the Peak-Over-Threshold / Point Poisson Process
(default thresh_pot = 10 units) not used by default!}

\item{thresh_hbev}{threshold for the hbev family models
(default thresh_pot = 0 units)}

\item{pot_frac}{if true, use frac_exc to determine POT model threshold
instead of the given thresh_pot (default TRUE)}

\item{frac_exc}{fraction of non-zero observations above threshold
(default is 0.05, for POT model only)}

\item{iter}{number of iterations in each chains for HMC.
By default 50% burn it period.}

\item{chains}{number of parallel chains for HMC sampler}

\item{onlymax_gev}{if TRUE, allow for data = vector of annual maxima
(default FALSE, only for GEV model)}

\item{refresh}{To show output or not (default refresh = 0, no output shown)}

\item{empirical_prior}{If TRUE, use some empirical priors
(default empirical_prior = FALSE)}

\item{adapt_delta}{for HMC sampler (default adapt_delta = 0.8)}

\item{priorpar}{to provide specific prior parameter values (default NULL)}

\item{draw_priors}{if TRUE produce density for the priors
(mostly for plotting)}
}
\value{
a list with the following quantities:
\describe{
  \item{model}{model name}
  \item{Nt}{number of observations / block}
  \item{prior}{prior distributions used (parameters, type and possibly rng)}
  \item{Mgen}{number of samples for latent level paramaters}
  \item{thresh_pot}{threshold used for POT model}
 \item{thresh_hbev}{threshold used for hbev models}
 \item{model_fit}{Stan object with the model fit}
}
}
\description{
Fit an extreme value model to data
}
