% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbev_module.R
\name{split_obs_data}
\alias{split_obs_data}
\title{Splits a dataset in two parts to be used for calibration and testing.
  Extract two datasets for validation and calibration,
 with sizes M_val and M_cal years respectively.
 If M_cal and M_val are longer than the available
 time series, all available years are used instead.}
\usage{
split_obs_data(
  df,
  M_cal = 20,
  M_val = 200,
  Nt = 366,
  cross_val = TRUE,
  reshuffle = FALSE,
  flip_time = TRUE,
  reshuffle_days = FALSE,
  decluster = FALSE,
  signif_lim = 0.05
)
}
\arguments{
\item{df}{A data frame with the precipitation data. It must be a data frame
with fields PRCP (daily precipitation) and
YEAR (integer in format YYYY)}

\item{M_cal}{Number of years in the calibration dataset (for model fitting)}

\item{M_val}{Number of years in the validation dataset (for model testing)}

\item{Nt}{The number of events/year, or block size (default is Nt = 366).}

\item{cross_val}{If TRUE, the two datasets extracted are independent,
i.e, are extracted from different parts of the time series.
(default is TRUE). If FALSE, the two datasets can
(and in general do) contain
data from the same blocks of the original time series.}

\item{reshuffle}{If TRUE, resample randomly all years in the time series
before splitting in two dataset}

\item{flip_time}{If TRUE, flip the time series before splitting it
so that the calibration is extracted from the most recent values
and validation from earlier recorded values.
If false, do not flip time, so calibration and validation
samples are extracted in cronological order
(default is TRUE)}

\item{reshuffle_days}{If TRUE, resample randomly all daily values in the
original time series (default is FALSE)}

\item{decluster}{If true, apply declustering to the original time series
(default is FALSE)}

\item{signif_lim}{Correlation value used as threshold in the declustering
            (default signif_lim = 0.05).
            It is the value of autocorrelation below which
the time series is assumed to be uncorretated (used to determine the
window size for declustering if decluster = TRUE)}
}
\value{
Return a list of two lists, named datacal and dataval.
        Each list includes the following quantities:
\describe{
  \item{data}{matrix of size nyears*Nt with the daily values.
    Each row is an yearly sample (with non-zero values at the beginning)}
 \item{N}{yearly number of events}
 \item{max}{block (annual) maxima}
 \item{Xi}{sorted maxima}
 \item{Fi}{empirical non exceedanc frequency of the maxima Xi}
 \item{Tr}{empirical return time of the maxima Xi}
 \item{years}{years from the record used for the dataset}
 \item{nyears}{length of the dataset in years}
}
}
\description{
If M_cal or M_val are longer than the sample size,
 only the available data is used, thus producing
 shorter samples, except in the case
 of cross_validation when is not possible to produce an
 independent validation sample as expected
 (e.g., if M_cal > sample size an independent sample is not
 created and an error is raised.)
}
\details{
Id decluster = TRUE, before fitting the models we apply a
 declustering technique (default is FALSE):
 The time lag Tau at which the serial correlation decays below
 a given value 'signif_min' (default 0.05) is computed.
 Then the time series is declustered with a running window of
 length tau, in which only the largest observations is kept.

 signif_lim = Value of autocorrelation below which
 the time series is assumed to be uncorretated (used to determine the
 window size for declustering if decluster = TRUE)

 If reshuffle -> the order of the years in the record
 is reshuffled before splitting the series in the two datasets.
 (default id FALSE)

 If reshuffle_days -> the number of events and the intensities
                      of all events in the time series are
                      resampled without resubstitution
                      (default is FALSE)

 If flip_time -> If not reshuffle, start extracting years
                 from the end of the record,
                 flipping the order of observed yerars
                 (e.g., to simulate longer time series that
                 go progressively back longer in time)
                 extract as calibration the last M_cal years,
                 (beacuse more recent data may have better resolution),
                 and then extract the M_val years before (if cross_val),
                 or again extract the last M_val years (with overlapping)
                 if cross_val is FALSE.

 If neither reshuffle or flip time, start extracting
 samples from the beginning of the time series.
}
