% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbev_module.R
\name{table_max}
\alias{table_max}
\title{Compute yearly rainfall statistics from a time series}
\usage{
table_max(df, Nt = 366, reshuffle_days = FALSE)
}
\arguments{
\item{df}{A data frame with the data.}

\item{Nt}{= 366 The number of events / year.}

\item{reshuffle_days}{= FALSE wether or not reshuffle the days.}
}
\value{
a list with various statistics.
}
\description{
Compute yearly rainfall statistics from a time series
}
