% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbev_module.R
\name{load_obs_data}
\alias{load_obs_data}
\title{Load a dataset of rainfall observations}
\usage{
load_obs_data(
  filepath,
  readdata = TRUE,
  maxmiss = 30,
  min_nevents = 0,
  dividebyten = TRUE,
  Nt = 366
)
}
\value{
df, dataframe with the precipitation data
}
\description{
Load a dataset of rainfall observations
}
